refactor: move types and helper methods
