+++ packages/client/src/encryption/GroupKeyStoreFactory.ts
@@ -1,16 +1,16 @@
 import { scoped, Lifecycle, inject } from 'tsyringe'
-
 import { instanceId } from '../utils/utils'
 import { CacheAsyncFn } from '../utils/caches'
 import { inspect } from '../utils/log'
 import { Context, ContextError } from '../utils/Context'
-import { ConfigInjectionToken, CacheConfig } from '../Config'
-
-import { EncryptionConfig, GroupKeysSerialized, parseGroupKeys } from './KeyExchangeStream'
+import { ConfigInjectionToken, CacheConfig, EncryptionConfig } from '../Config'
+import { GroupKeyId, GroupKeyish } from './GroupKey'
 import { GroupKeyStore } from './GroupKeyStore'
 import { GroupKey } from './GroupKey'
 import { StreamID } from 'streamr-client-protocol'
 import { Authentication, AuthenticationInjectionToken } from '../Authentication'
+
+export type GroupKeysSerialized = Record<GroupKeyId, GroupKeyish>
 
 // In the client API we use the term EncryptionKey instead of GroupKey.
 // The GroupKey name comes from the protocol. TODO: we could rename all classes
@@ -20,6 +20,13 @@ export interface UpdateEncryptionKeyOptions {
     streamId: string
     distributionMethod: 'rotate' | 'rekey'
     key?: GroupKey
+}
+
+function parseGroupKeys(groupKeys: GroupKeysSerialized = {}): Map<GroupKeyId, GroupKey> {
+    return new Map<GroupKeyId, GroupKey>(Object.entries(groupKeys || {}).map(([key, value]) => {
+        if (!value || !key) { return null }
+        return [key, GroupKey.from(value)]
+    }).filter(Boolean) as [])
 }
 
 @scoped(Lifecycle.ContainerScoped)
